% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{conversion_rates}
\alias{conversion_rates}
\title{Get Current Conversion Rates}
\usage{
conversion_rates(currency = "USD", key, ...)
}
\arguments{
\item{currency}{The currency to use as a baseline. By default this is US Dollars; while it can
be changed to other currencies, that does require an 'Enterprise' OER key, and will otherwise
fail.}

\item{key}{Your Open Exchange Rates API key. See their \href{https://openexchangerates.org/signup}{access plans}
(particularly the 'forever free' plan linked at the bottom).}

\item{...}{Further arguments to pass to httr's \code{GET} function.}
}
\description{
Get the current conversion rates between various different currencies.
}

