% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{currency_convert}
\alias{currency_convert}
\title{Convert between currencies using the Open Exchange Rates API}
\usage{
currency_convert(x, from = "USD", to, key, ...)
}
\arguments{
\item{x}{a vector of amounts, formatted as integer or numeric values.}

\item{from}{the currency \code{x} is currently in. US Dollars by default.}

\item{to}{the currency you \emph{want} \code{x} to be in.}

\item{key}{your Open Exchange Rates API key. See their \href{https://openexchangerates.org/signup}{access plans}
(particularly the 'forever free' plan linked at the bottom).}

\item{...}{further arguments to pass to httr's \code{GET} function.}
}
\description{
Uses the Open Exchange Rates API to convert a series of amounts between
two currencies.
}

