% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reformat.R
\name{from_currency}
\alias{from_currency}
\title{Convert currency-formatted strings into numeric values}
\usage{
from_currency(x, decimal_delim = ".")
}
\arguments{
\item{x}{a vector of strings, formatted as money amounts. See \code{\link{to_currency}}.}

\item{decimal_delim}{the character used to delimit the decimal amount. Set to "." by default.}
}
\description{
takes a vector of strings formatted as amounts of money ("$12,329.34")
and reformats them as numeric values (12329.34).
}
\examples{
from_currency("£1,249.34")
# [1] 1249.34
}

