% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{historic_currency}
\alias{historic_currency}
\title{Get Historic Currency Data}
\usage{
historic_currency(dates, currency = "USD", key, ...)
}
\arguments{
\item{dates}{a vector of dates to get data for. Strings or date objects.}

\item{currency}{the currency to use as a baseline. By default this is US Dollars; while it can
be changed to other currencies, that does require an 'Enterprise' OER key, and will otherwise
fail.}

\item{key}{Your Open Exchange Rates API key. See their \href{https://openexchangerates.org/signup}{access plans}
(particularly the 'forever free' plan linked at the bottom).}

\item{...}{Further arguments to pass to httr's \code{GET} function.}
}
\description{
Get Historic Currency Data
}

