% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{list_currencies}
\alias{list_currencies}
\title{Find Supported Currencies}
\usage{
list_currencies(key, as_df = FALSE, ...)
}
\arguments{
\item{key}{Your Open Exchange Rates API key. See their \href{https://openexchangerates.org/signup}{access plans}
(particularly the 'forever free' plan linked at the bottom).}

\item{as_df}{Whether to return the results as a data.frame. If not, it will be a named list. Set to FALSE
by default.}

\item{...}{Further arguments to pass to httr's \code{GET} function.}
}
\description{
Generates a list of the currencies supported by \code{\link{currency_convert}}.
}

