% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lucr.R
\docType{package}
\name{lucr}
\alias{lucr}
\alias{lucr-package}
\title{Currency Formatting and Conversion}
\description{
Reformat currency-based data as numeric values (or numeric values as currency-based data)
and convert between currencies.
}

