% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrency.R
\docType{package}
\name{rrency}
\alias{rrency}
\alias{rrency-package}
\title{Currency Formatting and Conversion}
\description{
Reformat currency-based data as numeric values (or numeric values as currency-based data)
and convert between currencies.
}

